Agregado modulo deudas, con errores conocidos (el modelo articulo crea un bucle infinito en la llamada)

+++ public/deudas/main.js
@@ -1,13 +1,15 @@
-
 new Vue({
     el : '#deudas',
     data: {
-        articulo : null,
+        deuda : { },
+        articulo : {marca : {}},
+        selected : 0,
         descuento: 0,
         cantidad : 1,
         interes : 0,
         deuda_id : 0,
         estado : null,
+        estado_real : null,
         detalles_subtotal : 0,
         importe_total : 0,
         articulos : [],
@@ -21,25 +23,52 @@ new Vue({
                     subtotal : ''}
     },
     methods : {
+
+        borrarArt(index)
+        {
+            Swal({
+                title: 'Estas segudo de eliminar este articulo?',
+                text: 'Este articulo ya no se volvera a recuperar!',
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonText: 'Si, borralo!',
+                cancelButtonText: 'No, mantenlo'
+            }).then((result) => {
+                if (result.value) {
+                    url = '/api/detalle_deudas/' + index;
+                    axios.delete(url).then(response => {
+                        this.detalles = this.obtenerDetalles();
+                        this.actualizarSubtotal();
+                        Swal(
+                            'Eliminado!',
+                            'El articulo fue eliminado.',
+                            'success'
+                        )
+                    }).catch(error => {
+                        console.log(error.response.data.data.message);
+                    });
+
+                    // For more information about handling dismissals please visit
+                    // https://sweetalert2.github.io/#handling-dismissals
+                } else if (result.dismiss === Swal.DismissReason.cancel) {
+                    Swal(
+                        'Cancelado',
+                        'El articulo no fue eliminado!',
+                        'error'
+                    )
+                }
+            })
+        },
+
         obtenerDetalles()
         {
-            url = '/detalle_deudas/deuda/' + this.deuda_id;
+            var url = '/detalle_deudas/deuda/' + this.deuda_id;
+            var self = this;
             axios.get(url).then(response => {
-                this.detalles = response.data.data;
+                self.detalles = response.data.data;
             }).catch(error => {
                 console.log(error.response.data);
             });
-        },
-        borrarArticulo(index){
-            url = '/api/detalle_deudas/' + index;
-            axios.delete(url).then(response => {
-                toastr.warning('Articulo eliminado correctamente!');
-                this.detalles = this.obtenerDetalles();
-                this.actualizarSubtotal();
-            }).catch(error => {
-                console.log(error.response.data.message);
-            });
-            
         },
         agregarArticulo(){
             if(this.articulo != null){
@@ -52,13 +81,22 @@ new Vue({
                 this.detalle.descuento = this.descuento;
                 this.detalle.subtotal = this.subtotal(this.articulo.precio_venta, this.descuento, this.cantidad);
                 axios.post(url, this.detalle).then(response => {
+                    this.articulo = {marca: {}};
+                    this.selected = 0;
                     this.detalles = this.obtenerDetalles();
                     this.actualizarSubtotal();
+                    this.actualizarImporteTotal();
                     this.descuento = 0,
                     this.cantidad = 1,
-                    toastr.success('Detalle guardado exitosamente!');
+                    swal({
+                        position: 'top-end',
+                        type: 'success',
+                        title: 'Articulo agregado correctamente',
+                        showConfirmButton: false,
+                        timer: 1500
+                    })
                 }).catch(error => {
-                    toastr.error(error.response.data.message);
+                    toastr.error(error);
                 });
             }
             else{
@@ -75,34 +113,113 @@ new Vue({
         },
         obtenerArticulos()
         {
+
+            var articulo = { id : 0, nombre: 'Seleccione un articulo'};
             axios.get('/api/articulos').then(response => {
-                this.articulos = response.data.data;
+                var articulos = response.data.data;
+                articulos.unshift(articulo);
+                this.articulos = articulos;
+            }).catch(error => {
+                //toastr.error(error.response.data.message);
+
             });
         },
         actualizarSubtotal()
         {
-            url = '/admin/deudas/importe/' + this.deuda_id;
+            url = '/admin/deudas/subtotal/' + this.deuda_id;
             axios.get(url).then(response => {
                 this.detalles_subtotal = response.data;
+                this.actualizarImporteTotal(response.data);
             }).catch(error => {
                 console.log(error.response.data.message);
             });
+        },
+        actualizarImporteTotal(subtotal)
+        {
+            url = '/admin/deudas/total/' + this.deuda_id;
+            axios.get(url).then(response => {
+                this.importe_total = response.data;
+            })
+
+
         },
         obtenerDeuda()
         {
             url = '/api/deudas/'+ this.deuda_id;
             axios.get(url).then(response => {
                 this.estado = response.data.data.estado;
-                this.interes = response.data.data.interes;
+                response.data.data.interes ? this.interes = response.data.data.interes : 0 ;
                 this.importe_total = response.data.data.importe_total;
+                this.deuda = response.data.data;
+            });
+        },
+        obtenerArticulo()
+        {
+            if(this.selected != 0)
+            {
+                let self = this;
+                url = '/api/articulos/'+ this.selected;
+
+                axios.get(url).then(response => {
+                    let articulo = response.data.data;
+                    self.articulo = articulo;
+                });
+            }
+        },
+        validaCantidad()
+        {
+            if(this.cantidad < 1 )
+            {
+                this.cantidad = 1;
+            }
+            if(this.cantidad > this.articulo.cantidad)
+            {
+                this.cantidad = this.articulo.cantidad;
+            }
+
+        },
+        validaDescuento()
+        {
+            if(this.descuento < 0)
+            {
+                this.descuento = 0;
+            }
+            if(this.descuento > 100)
+            {
+
+                this.descuento = 100;
+            }
+        },
+        validaInteres()
+        {
+            if(this.interes < 0)
+            {
+                this.interes = 0;
+            }
+        },
+        simular()
+        {
+            var self = this;
+            var interes = parseFloat(self.interes*0.01).toFixed(2);
+            var importe_interes =interes*self.detalles_subtotal;
+            var importe_total = importe_interes + self.detalles_subtotal;
+
+            Swal({
+                type: 'info',
+                title: 'Nuevo importe: $' + importe_total,
+                text: 'Se refleja el importe subtotal + $' + importe_interes + ' de intereses!',
+                customClass: 'swal-wide',
+
             });
         }
     },
     mounted(){
             this.deuda_id = document.querySelector("#deuda_id").value;
+            this.estado_real = document.querySelector("#estado_real").value;
             this.obtenerDetalles();
             this.obtenerArticulos();
             this.actualizarSubtotal();
             this.obtenerDeuda();
+
         }
 });
